requirements.txt
Flask==2.3.3
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0